y_pred was passed as int?

+++ benchmark/wifi_csi/train.py
@@ -101,7 +101,7 @@ def train(model: Module,
         data_batch_y = data_batch_y.detach().cpu().numpy()
         predict_train_y = predict_train_y.detach().cpu().numpy()
 
-        dict_error_train = performance_metrics(data_batch_y.astype(int), predict_train_y.astype(int),
+        dict_error_train = performance_metrics(data_batch_y.astype(int), predict_train_y,
                                                var_mode=var_mode, var_threshold=var_threshold)
 
         model.eval()
